services/auth_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4
httpx==0.27.2
pika==1.3.2
structlog==24.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
asyncpg
email-validator
python-multipart==0.0.9


